+++ src/azure-cli/HISTORY.rst
@@ -1620,6 +1620,7 @@ Release History
 
 **Redis**
 
+* Add preview param `--ports` to `az redis force-reboot` command
 * Add preview param `--replicas-per-master` to `az redis create` command
 * Update azure-mgmt-redis from 6.0.0 to 7.0.0rc1